+++ empty_single_square.py
@@ -1,0 +1,7 @@
+import single_square
+
+
+class EmptySingleSquare(single_square.SingleSquare):
+    def __init__(self, parent):
+        super(EmptySingleSquare, self).__init__(parent, "white")
+        print ("do empty square stuff in here")
