+++ walking_square.py
@@ -1,0 +1,6 @@
+from single_square import SingleSquare
+
+
+class WalkingSquare(SingleSquare):
+    def __init__(self, parent):
+        super(WalkingSquare, self).__init__(parent, "yellow")
